Load blog post summaries directly from utils in the route loader.
