--- apps/blog/rpc/index.ts
@@ -1,1 +1,0 @@
-export * from "./get-blog-post-summaries"
