+++ apps/blog/utils.test.ts
@@ -3,6 +3,7 @@ import type { BlogPostFrontmatter, BlogPostMdxModule } from "./types"
 import {
   getBlogPostFrontmatter,
   getBlogPostModuleBySlug,
+  getBlogPostSummaries,
   toBlogPostFrontmatter,
 } from "./utils"
 
@@ -116,5 +117,163 @@ describe("getBlogPostModuleBySlug", () => {
     ])
 
     expect(getBlogPostModuleBySlug(modules, "unknown")).toBeNull()
+  })
+})
+
+describe("getBlogPostSummaries", () => {
+  it("includes keywords when provided in frontmatter", () => {
+    const modules = Object.fromEntries([
+      mdx(
+        "./content/tagged.mdx",
+        validFrontmatter({
+          title: "Tagged post",
+          keywords: ["react", "cloudflare"],
+        }),
+      ),
+    ])
+
+    const posts = getBlogPostSummaries(modules)
+    expect(posts[0]?.keywords).toEqual(["react", "cloudflare"])
+  })
+
+  it("maps frontmatter to summaries with slug from path", () => {
+    const modules = Object.fromEntries([
+      mdx(
+        "./content/hello-world.mdx",
+        validFrontmatter({
+          title: "Hello world",
+          description: "A short intro",
+          date: "2026-05-18",
+          imgSrc: "https://example.com/cover.jpg",
+          author: { name: "Jane Doe", role: "Editor" },
+        }),
+      ),
+    ])
+
+    expect(getBlogPostSummaries(modules)).toEqual([
+      {
+        slug: "hello-world",
+        title: "Hello world",
+        description: "A short intro",
+        date: "2026-05-18",
+        imgSrc: "https://example.com/cover.jpg",
+        author: { name: "Jane Doe", role: "Editor" },
+      },
+    ])
+  })
+
+  it("extracts slug from Windows-style paths", () => {
+    const modules = Object.fromEntries([
+      mdx(
+        "apps\\blog\\content\\win-post.mdx",
+        validFrontmatter({
+          title: "Windows path",
+          description: "Desc",
+          date: "2026-01-01",
+          imgSrc: "https://example.com/a.jpg",
+        }),
+      ),
+    ])
+
+    expect(getBlogPostSummaries(modules)[0]?.slug).toBe("win-post")
+  })
+
+  it("skips modules without a parsable slug", () => {
+    const modules = {
+      "invalid-path": {
+        frontmatter: validFrontmatter({ title: "Orphan" }),
+        default: () => null,
+      },
+    }
+
+    expect(getBlogPostSummaries(modules)).toEqual([])
+  })
+
+  it("throws when frontmatter is missing required fields", () => {
+    const modules = {
+      ...Object.fromEntries([
+        mdx(
+          "./content/no-title.mdx",
+          validFrontmatter({
+            title: "",
+            description: "Desc",
+            date: "2026-01-01",
+            author: { name: "Author" },
+          }),
+        ),
+      ]),
+      "./content/missing-title.mdx": { default: () => null },
+    }
+
+    expect(() => getBlogPostSummaries(modules)).toThrow()
+  })
+
+  it("throws when required frontmatter fields are missing", () => {
+    const modules = Object.fromEntries([
+      mdx("./content/minimal.mdx", {
+        title: "Minimal",
+      } as BlogPostFrontmatter),
+    ])
+
+    expect(() => getBlogPostSummaries(modules)).toThrow()
+  })
+
+  it("uses imgSrc from MDX frontmatter as-is", () => {
+    const modules = Object.fromEntries([
+      mdx(
+        "./content/local-cover.mdx",
+        validFrontmatter({
+          title: "Local cover",
+          imgSrc: "/assets/test-cover.png",
+        }),
+      ),
+    ])
+
+    expect(getBlogPostSummaries(modules)[0]?.imgSrc).toBe(
+      "/assets/test-cover.png",
+    )
+  })
+
+  it("throws when date is not a valid ISO date", () => {
+    const modules = {
+      "./content/bad-date.mdx": {
+        frontmatter: {
+          title: "Bad date",
+          description: "Desc",
+          date: "not-a-date",
+          imgSrc: "/assets/test-cover.png",
+          author: { name: "Author" },
+        } as BlogPostFrontmatter,
+        default: () => null,
+      },
+    }
+
+    expect(() => getBlogPostSummaries(modules)).toThrow()
+  })
+
+  it("sorts posts by date descending", () => {
+    const modules = Object.fromEntries([
+      mdx(
+        "./content/older.mdx",
+        validFrontmatter({
+          title: "Older",
+          date: "2024-01-01",
+          imgSrc: "https://example.com/a.jpg",
+        }),
+      ),
+      mdx(
+        "./content/newer.mdx",
+        validFrontmatter({
+          title: "Newer",
+          date: "2026-06-01",
+          imgSrc: "https://example.com/b.jpg",
+        }),
+      ),
+    ])
+
+    expect(getBlogPostSummaries(modules).map((post) => post.slug)).toEqual([
+      "newer",
+      "older",
+    ])
   })
 })
